@@ -1,22 +1,19 @@
-﻿using System;
+﻿namespace builtin_types;
 
-namespace builtin_types
+public static class BoolType
 {
-    public static class BoolType
+    public static void Examples()
     {
-        public static void Examples()
-        {
-            Literals();
-        }
+        Literals();
+    }
 
-        private static void Literals()
-        {
-            // <SnippetLiterals>
-            bool check = true;
-            Console.WriteLine(check ? "Checked" : "Not checked");  // output: Checked
+    private static void Literals()
+    {
+        // <SnippetLiterals>
+        bool check = true;
+        Console.WriteLine(check ? "Checked" : "Not checked");  // output: Checked
 
-            Console.WriteLine(false ? "Checked" : "Not checked");  // output: Not checked
-            // </SnippetLiterals>
-        }
+        Console.WriteLine(false ? "Checked" : "Not checked");  // output: Not checked
+        // </SnippetLiterals>
     }
 }
