@@ -9,7 +9,7 @@
             .WriteRequestToConsole();
 
         var jsonResponse = await response.Content.ReadAsStringAsync();
-        WriteLine($"{jsonResponse}\n");
+        Console.WriteLine($"{jsonResponse}\n");
 
         // Expected output
         //   DELETE https://jsonplaceholder.typicode.com/todos/1 HTTP/1.1
