@@ -9,10 +9,10 @@
             .WriteRequestToConsole();
         
         var jsonResponse = await response.Content.ReadAsStringAsync();
-        WriteLine($"{jsonResponse}\n");
+        Console.WriteLine($"{jsonResponse}\n");
 
         // Expected output:
-        //   GET https://jsonplaceholder.typicode.com/todos/3 HTTP/ 1.1
+        //   GET https://jsonplaceholder.typicode.com/todos/3 HTTP/1.1
         //   {
         //     "userId": 1,
         //     "id": 3,
