@@ -6,9 +6,9 @@
         var todos = await httpClient.GetFromJsonAsync<List<Todo>>(
             "todos?userId=1&completed=false");
 
-        WriteLine("GET https://jsonplaceholder.typicode.com/todos?userId=1&completed=false HTTP/1.1");
-        todos?.ForEach(WriteLine);
-        WriteLine();
+        Console.WriteLine("GET https://jsonplaceholder.typicode.com/todos?userId=1&completed=false HTTP/1.1");
+        todos?.ForEach(Console.WriteLine);
+        Console.WriteLine();
 
         // Expected output:
         //   GET https://jsonplaceholder.typicode.com/todos?userId=1&completed=false HTTP/1.1
