@@ -1,4 +1,6 @@
-﻿using System;
+using System;
+
+namespace ExprBodied;
 
 // <Snippet1>
 public class Location
